+++ Packages/UGF.EditorTools/Editor/Assets/AssetIdEditorUtility.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: b72caa76ff614cb3a0c2d5d0d78a4212
+timeCreated: 1687447798
